+++ pcsd/settings.rb
@@ -25,6 +25,7 @@ COROSYNC_QDEVICE_NET_SERVER_CA_FILE = (
   COROSYNC_QDEVICE_NET_SERVER_CERTS_DIR + "/qnetd-cacert.crt"
 )
 COROSYNC_QDEVICE_NET_CLIENT_CERTS_DIR = "/etc/corosync/qdevice/net/nssdb"
+COROSYNC_CONFIG_DIR = "/etc/corosync"
 
 SUPERUSER = 'hacluster'
 ADMIN_GROUP = 'haclient'