+++ CHANGELOG.md
@@ -33,6 +33,10 @@
   ([rhbz#1373614])
 - Removal of a pacemaker remote resource no longer causes the respective remote
   node to be fenced ([rhbz#1390609])
+  
+### Changed
+- `pcs cluster setup` configure corosync with an encrypted communication
+  ([rhbz#1165821])
 
 [ghissue#134]: https://github.com/ClusterLabs/pcs/issues/134
 [rhbz#1176018]: https://bugzilla.redhat.com/show_bug.cgi?id=1176018
@@ -44,6 +48,7 @@
 [rhbz#1390609]: https://bugzilla.redhat.com/show_bug.cgi?id=1390609
 [rhbz#1422667]: https://bugzilla.redhat.com/show_bug.cgi?id=1422667
 [rhbz#1433016]: https://bugzilla.redhat.com/show_bug.cgi?id=1433016
+[rhbz#1165821]: https://bugzilla.redhat.com/show_bug.cgi?id=1165821
 
 
 ## [0.9.157] - 2017-04-10